app/requirements.txt
numpy==1.22.3
openpyxl==3.0.9
pandas==1.4.2
psycopg2-binary==2.9.3
pyarrow==7.0.0
pytz==2022.1
requests==2.27.1
SQLAlchemy==1.4.35
urllib3==1.26.9
httplib2==0.20.4